requirements.txt
torch>=2.6.0
numpy>=2.0
scipy>=1.14
matplotlib>=3.9
pyyaml>=6.0
tqdm>=4.66
networkx>=3.2
scikit-learn>=1.4
basemap>=2.0
tigramite>=5.2
